my-server/requirements.txt
## The following requirements were added by pip freeze:
aniso8601==8.0.0
astroid==2.4.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
dataclasses-json==0.4.2
dnspython==1.16.0
eventlet==0.27.0
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SocketIO==4.3.1
google-api-core==1.22.2
google-auth==1.21.1
google-cloud-core==1.4.1
google-cloud-storage==1.31.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
greenlet==0.4.16
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.6.0
marshmallow-enum==1.5.1
monotonic==1.5
mypy-extensions==0.4.3
protobuf==3.13.0
py-ts-interfaces==0.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-engineio==3.12.1
python-socketio==4.5.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
rsa==4.6
six==1.14.0
stringcase==1.2.0
typing-extensions==3.7.4.2
typing-inspect==0.6.0
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zope.event==4.4
zope.interface==5.1.0
